criado modelo e controler de Categoria de Certificado
